home: Fix integrations count.
