+++ templates/corporate/hello.html
@@ -390,7 +390,7 @@
                 <h1>Seamless integrations with everything you use.</h1>
             </header>
             <p>
-                Zulip has more than 90 native integrations. Several hundred more
+                Zulip has more than {{ integrations_count_display }} native integrations. Several hundred more
                 are available through
                 <a href="/integrations/doc/zapier">Zapier</a>
                 and
